Update Documentation for Public Cloud service
If LogDNA service is running in one of the public providers the current documentation is not clear how to configure this plugin to work against such service. Consumer of this plugin would have to go through the plugin code to understand the default value and advanced option they can set to make it work.

This PR will help improve usability of this Plugin.